[KOGITO-877] - Improve the KOGITO images push-stating script to also increase the tag on image-stream
